[Feature] DRC-357 Add preset checks into recces's state file
